feat: LobbyLayout -> LobbyPage로 통합
