+++ src/pages/lobby/LobbyPage.jsx
@@ -1,4 +1,8 @@
-import useGameSocket from "hooks/useGameSocket";
+import Lobby from "components/lobby/Lobby";
+import { ROOT_URL } from "constants/url";
+import { tokenReissueAPI } from "services/auth/auth";
+import "styles/pages/lobby/LobbyPage.scss";
+import { getAuthToken } from "utils/token";
 
 const LobbyPage = () => {
   // 소켓 연결 테스트(지워도 됨!)
@@ -7,7 +11,30 @@ const LobbyPage = () => {
 
   useGameSocket({ nickName, roomId });
 
-  return <div className="robby-page-container">lobby</div>;
+  return (
+    <div className="lobby-page-container">
+      <Lobby />
+    </div>
+  );
 };
 
 export default LobbyPage;
+
+export const loader = async () => {
+  const { accessToken, setAccessToken } = getAuthToken();
+
+  if (!accessToken) {
+    // 토큰 재발급
+    try {
+      const res = await tokenReissueAPI();
+      const newAccessToken = res.accessToken;
+      const newExpireTime = res.expiredAt;
+
+      setAccessToken(newAccessToken, newExpireTime);
+    } catch (e) {
+      return redirect(ROOT_URL);
+    }
+  }
+
+  return;
+};
